GH-1138 Keep hellers in the account breakdown
The infopult screen returned 500 for any participant whose payment was
not a whole number: the transaction split declared int while the amounts
it is built from are float, and PHP 8 refuses the implicit conversion.
97 payments this year have a fractional amount, across 41 participants.

Had it not thrown, the hellers would have been cut off instead and the
balance would have been wrong by that much, so the type is corrected
along the whole chain rather than at the entry point.

Sums are rounded where they are taken, because adding decimal amounts
leaves a residue around 1e-15 and the infopult reads a negative total as
an unpaid balance. Amounts are formatted where they are rendered,
because negating a float zero prints -0 and most participants have
categories they ordered nothing from.

+++ model/Accounting/TransactionSplit.php
@@ -7,18 +7,19 @@ namespace Gamecon\Accounting;
 class TransactionSplit
 {
     /**
-     * Positive amount means INCREASE in USER's balance, negative decrease
+     * Positive amount means INCREASE in USER's balance, negative decrease.
+     * Float, because payments arrive from the bank in hellers.
      */
-    private int $amount;
+    private float $amount;
     private string $description;
 
-    public function __construct(int $amount, string $description)
+    public function __construct(float $amount, string $description)
     {
         $this->amount = $amount;
         $this->description = $description;
     }
 
-    public function getAmount(): int
+    public function getAmount(): float
     {
         return $this->amount;
     }
